Add finally block to close redis connections

+++ main.py
@@ -234,12 +234,12 @@ def process_audio(input_path, chat_id):
         else:
             bot.send_message(chat_id, "Failed to compress audio.")
     except Exception as e:
-        print(f"Unexpected error: {e}")
+        logger.error(f"Unexpected error: {e}")
 
     finally:
-        if input_path:
+        if input_path and os.path.exists(input_path):
             os.remove(input_path)
-        if output_path:
+        if output_path and os.path.exists(output_path):
             os.remove(output_path)
 
 
@@ -247,12 +247,15 @@ def prompt_for_email_option(chat_id, report):
     """Prompt the user for email options."""
     report_id = str(uuid.uuid4())
     redis_client = Redis(connection_pool=pool)
-    redis_client.set(f"message:{report_id}", report, ex=6*60*60)
-
-    current_datetime = datetime.datetime.now(tz=pytz.utc)
-    formatted_date = current_datetime.strftime("%d/%m/%Y")
-    redis_client.set(f"subject:{report_id}", f"Notes {formatted_date}", ex=6*60*60)
-    redis_client.close()
+    try:
+        redis_client.set(f"message:{report_id}", report, ex=6*60*60)
+        current_datetime = datetime.datetime.now(tz=pytz.utc)
+        formatted_date = current_datetime.strftime("%d/%m/%Y")
+        redis_client.set(f"subject:{report_id}", f"Notes {formatted_date}", ex=6*60*60)
+    except Exception as e:
+        logger.error(f"Error setting data in Redis: {e}")
+    finally:
+        redis_client.close()
 
     markup = types.InlineKeyboardMarkup()
     markup.add(
@@ -304,14 +307,15 @@ def save_subject_with_logging(message, report_id):
 def save_subject(message, report_id):
     """Save the new subject."""
     logger.info(f"Saving new subject: {message.text} for report_id: {report_id}")
+    redis_client = Redis(connection_pool=pool)
     try:
-        redis_client = Redis(connection_pool=pool)
         redis_client.set(f"subject:{report_id}", message.text, ex=6*60*60)
         logger.info(f"Saved new subject: {message.text} for report_id: {report_id}")
-        redis_client.close()
         display_report(message.chat.id, report_id)
     except Exception as e:
         logger.error(f"Error saving subject to Redis: {e}")
+    finally:
+        redis_client.close()
 
 
 @bot.callback_query_handler(func=lambda call: call.data.startswith("edit_message"))
@@ -327,11 +331,14 @@ def handle_edit_message(call):
 def save_message(message, report_id):
     """Save the new message."""
     redis_client = Redis(connection_pool=pool)
-    redis_client.set(f"message:{report_id}", message.text, ex=6*60*60)
-    logger.info(f"Saved new message for report_id: {report_id}")
-    redis_client.close()
-
-    display_report(message.chat.id, report_id)
+    try:
+        redis_client.set(f"message:{report_id}", message.text, ex=6*60*60)
+        logger.info(f"Saved new message for report_id: {report_id}")
+        display_report(message.chat.id, report_id)
+    except Exception as e:
+        logger.error(f"Error saving message to Redis: {e}")
+    finally:
+        redis_client.close()
 
 
 @bot.callback_query_handler(func=lambda call: call.data.startswith("send_email"))
@@ -340,79 +347,84 @@ def handle_send_email(call):
     redis_client = Redis(connection_pool=pool)
     report_id = call.data.split(":", 1)[1]
 
-    subject = redis_client.get(f"subject:{report_id}")
-    message = redis_client.get(f"message:{report_id}")
+    try:
+        subject = redis_client.get(f"subject:{report_id}")
+        message = redis_client.get(f"message:{report_id}")
 
-    if subject:
-        subject = subject.decode("utf-8")
-    else:
-        current_datetime = datetime.datetime.now(tz=pytz.utc)
-        formatted_date = current_datetime.strftime("%d/%m/%Y")
+        if subject:
+            subject = subject.decode("utf-8")
+        else:
+            current_datetime = datetime.datetime.now(tz=pytz.utc)
+            formatted_date = current_datetime.strftime("%d/%m/%Y")
+            subject = f"Notes {formatted_date}"
 
-        subject = f"Notes {formatted_date}"
+        if message:
+            message = message.decode("utf-8")
+        else:
+            bot.send_message(call.message.chat.id, "Report not found.")
+            return
 
-    if message:
-        message = message.decode("utf-8")
-    else:
-        bot.send_message(call.message.chat.id, "Report not found.")
+        if TO_EMAIL:
+            for email in TO_EMAIL:
+                send_email(subject, message, email)
+
+        redis_client.delete(report_id)
+        redis_client.delete(f"subject:{report_id}")
+        redis_client.delete(f"message:{report_id}")
+        bot.send_message(call.message.chat.id, "Email sent successfully.")
+    except Exception as e:
+        logger.error(f"Error handling email sending: {e}")
+    finally:
         redis_client.close()
-        return
-
-    if TO_EMAIL:
-        for email in TO_EMAIL:
-            send_email(subject, message, email)
-
-    redis_client.delete(report_id)
-    redis_client.delete(f"subject:{report_id}")
-    redis_client.delete(f"message:{report_id}")
-    bot.send_message(call.message.chat.id, "Email sent successfully.")
-    redis_client.close()
 
 
 def display_report(chat_id, report_id):
     """Display the report with options to edit or send."""
     redis_client = Redis(connection_pool=pool)
-    subject = redis_client.get(f"subject:{report_id}")
-    message = redis_client.get(f"message:{report_id}")
+    try:
+        subject = redis_client.get(f"subject:{report_id}")
+        message = redis_client.get(f"message:{report_id}")
 
-    if subject:
-        subject = subject.decode("utf-8")
-    else:
-        subject = "(No Subject)"
+        if subject:
+            subject = subject.decode("utf-8")
+        else:
+            subject = "(No Subject)"
 
-    if message:
-        message = message.decode("utf-8")
-    else:
-        message = "(No Message)"
+        if message:
+            message = message.decode("utf-8")
+        else:
+            message = "(No Message)"
 
-    redis_client.close()
-
-    response = f"""
+        response = f"""
 Subject: 
 {subject}
     
 Body:
 {message}
-    """
+        """
 
-    markup = types.InlineKeyboardMarkup()
-    markup.add(
-        types.InlineKeyboardButton(
-            "Edit Subject", callback_data=f"edit_subject:{report_id}"
+        markup = types.InlineKeyboardMarkup()
+        markup.add(
+            types.InlineKeyboardButton(
+                "Edit Subject", callback_data=f"edit_subject:{report_id}"
+            )
         )
-    )
-    markup.add(
-        types.InlineKeyboardButton(
-            "Edit Body", callback_data=f"edit_message:{report_id}"
+        markup.add(
+            types.InlineKeyboardButton(
+                "Edit Body", callback_data=f"edit_message:{report_id}"
+            )
         )
-    )
-    markup.add(
-        types.InlineKeyboardButton(
-            "Send Email", callback_data=f"send_email:{report_id}"
+        markup.add(
+            types.InlineKeyboardButton(
+                "Send Email", callback_data=f"send_email:{report_id}"
+            )
         )
-    )
 
-    bot.send_message(chat_id, response, reply_markup=markup)
+        bot.send_message(chat_id, response, reply_markup=markup)
+    except Exception as e:
+        logger.error(f"Error displaying report: {e}")
+    finally:
+        redis_client.close()
 
 
 @bot.message_handler(commands=["start", "restart"])
@@ -422,15 +434,6 @@ def start(message):
         "Hi! Send me an audio file and I will generate a written SOAP note in English."
     )
     bot.send_message(message.chat.id, message_to_send, parse_mode="Markdown")
-
-
-# @bot.message_handler(func=lambda message: True)
-# def handle_random_message(message):
-#     """Handle random text messages."""
-#     default_message = (
-#         "I'm sorry, I can only process audio files. Please send me an audio file and I will generate a written SOAP note in English."
-#     )
-#     bot.send_message(message.chat.id, default_message, parse_mode="Markdown")
 
 
 @app.route(f"/{WEBHOOK_SECRET}", methods=["POST"])
